requirements.txt
torch>=2.1.0
transformers>=4.37.0
accelerate>=0.25.0
peft>=0.7.0
trl>=0.7.0
bitsandbytes>=0.41.0
datasets>=2.16.0
sentencepiece>=0.1.99
protobuf>=3.20.0
mlflow>=2.9.0
pyyaml>=6.0
scipy>=1.11.0
numpy>=1.26.0
